requirements.txt
coverage-badge==1.0.1
discord.py==1.7.2
pytest==6.2.4
pytest-cov==2.12.1
requests==2.25.1